+++ src/main/java/com/mw/cc/dto/Currency.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @JsonIgnoreProperties(ignoreUnknown = true)
 @Data
 @NoArgsConstructor
@@ -11,6 +13,6 @@ public class Currency {
 
   private String currency;
   private String code;
-  private Double mid;
+  private BigDecimal mid;
 
 }
